IsoLog/pyproject.toml
[project]
name = "isolog"
version = "0.1.0"
description = "Portable SIEM for isolated/air-gapped networks"
authors = [
    {name = "IsoLog Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.23",
    "whoosh>=2.7.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.2",
    "watchdog>=3.0.0",
    "scikit-learn>=1.3.2",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "reportlab>=4.0.7",
    "pysigma>=0.10.0",
    "ruamel.yaml>=0.18.5",
    "cryptography>=41.0.7",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
